PostTransaction autosplit function - based on the payload size...  Number of peers and number of transactions in payload - autoadjusts to payload size
Peer Height selection - additional function

Code improvements and clean up. Splitting main file into more submodules...
