+++ cmd/arkgopool/.gitignore
@@ -2,3 +2,4 @@
 *.db
 *.exe
 *.lock
+
