+++ cmd/arkgopool/log/.gitignore
@@ -1,0 +1,5 @@
+# ignore all files in this dir...
+*
+
+# ... except for this one.
+!.gitignore
