+++ cmd/arkgopool/settings/banner.txt
@@ -8,4 +8,4 @@
         _____/ (_)__  ____  / /_
        / ___/ / / _ \/ __ \/ __/
       / /__/ / /  __/ / / / /_
-      \___/_/_/\___/_/ /_/\__/  v0.3.0 by chris&jarunik
+      \___/_/_/\___/_/ /_/\__/  v0.6.0 by chris&jarunik
